requirements.txt
sanic==22.6.2
#diffusers==0.4.1
#git+https://github.com/huggingface/diffusers@v0.5.1
#git+https://github.com/HazyResearch/diffusers
transformers==4.22.2
scipy==1.9.3
requests_futures==1.0.0
numpy==1.23.4
scikit-image==0.19.3
accelerate==0.13.2
triton==2.0.0.dev20221105
ftfy==6.1.1
spacy==3.4.3
boto3==1.26.12